Switch prio test/test-wmf job from hhvm to php72
For test/test-wmf pipelines where only a single phpflavour exist
of a job, and the job has both hhvm and php72 variants in the
gate pipeline for the same repo, switch the test/test-wmf job
from hhvm to php72.

Update the 'php' pipeline to still contain all php jobs not
part of the 'test' pipeline by default (now hhvm).

Rationale:
  The new MediaWikiUnitTestCase class in MediaWiki core cannot
  enforce its restrictions under HHVM due to an HHVM-bug. In order
  to keep the primary job useful and coveraging all aspects switch
  it to PHP72 instead. This still upholds all the same intentions
  as before (it's a subset of the gate, and its a flavour we run
  WMF production traffic on). And will make it so that it continues
  to the be job that has the most tests/coverage (the new tests
  from MediaWikiUnitTestCase are disabled on HHVM).
  More details at Ic22075bb5e8.

It also prepares us for the php72 migration making progress in
production as hhvm slowly becomes the compat/alternate variant
instead of the primary.

Change-Id: I776e2af5c92e6e0d7dd06db67f450ebdcb9d3d81