phan: stop installing MediaWiki
To run the phan static analysis, we do not need a fully installed
MediaWiki.
We also do not inject dependencies of dependencies, thus the installer
would fail due to a missing requirement. For example WikimediaMessages
refers to code in skins/MinervaNeue which in turn requires
MobileFrontend, though that is not required by WikimediaMessages, the
installer would fail with:

 A dependency error was encountered while installing the extension
 "MinervaNeue": Could not find the registration file for the extension
 &#34;MobileFrontend&#34;

Skip MediaWiki installer entirely: quibble --skip-install

Update:
    mwext-php72-phan-docker
    mwext-php72-phan-seccheck-docker
    mwext-php72-phan-seccheck-docker-non-voting
    mwskin-php72-phan-docker
    mwskin-php72-phan-docker-non-voting
    mwskin-php72-phan-seccheck-docker
    mwskin-php72-phan-seccheck-docker-non-voting

Bug: T232413
Change-Id: I59fd380b19c2b3f54d4588040306690418009a0c