Filter out debug spam when validating zuul layout
zuul-server -l validates the layout files but it is also hardcoded with
a DEBUG log level. With the introduction of several new pipelines, the
generate output is roughly 50 MBytes.  Jenkins takes minutes to process
the output and overall as a human we only care about the actual ERROR or
failure.

Filter out any DEBUG/INFO messages send to stderr for the 'zuul' log
bucket.  That effectively strip out all the spam and make the job faster
as well as making the reason of the failure easier to find.

Bug: T232287
Change-Id: I3b9195a7d9158c3ffdaf25b25efad7ad9b34af87