mediawiki-core-doxygen: Remove --no-extensions option to mwdocgen.php
There are no extensions installed in this Jenkins job, so it is not
needed. Also, in a core change I am removing this option so that
extensions are excluded by default (for local development convenience).
We have never afaik published Doxygen results for core + extensions.
While we might use the script for extensions at some point,
those would use --file, which will still work after I.

Bug: T232104
Change-Id: I8df3258e5d339f878d1fd876da705bdec70d0917