Handle codehealth non concurrency at Zuul level
Due to a limitation of the Jenkins Gearman plugin, non concurrenct
builds being hold by Jenkins QueueManager would cause all executors on
the node they get allocated to to be locked.  That prevent any other
jobs builds from happening on that Node which starve resources.

So that when we have ten queued build of
mwext-codehealth-master-non-voting , ten nodes are idling doing nothing!

Make the non concurrent codehealth jobs concurrent. Use a mutex at Zuul
level to prevent it from triggering the Gearman function immediately
which in turn mean the job is not added to the Jenkins Queue and the
nodes are not locking all their executors.

Update:
    mwcore-codehealth-master-non-voting
    mwext-codehealth-master-non-voting

Bug: T234259
Change-Id: I414f1b0c84c132f8fc3d2e37f9ff7834142cd54a