docker: upgrade setuptools in Quibble images
You shall never use the Debian setuptools package since it is quickly
outdated and missing features.

quibble 0.0.36 relies on setuptools_scm to detect the software version
from git and the module has to be listed under setup_requires.

That is not recognized in setup.cfg until setuptools 36.7.0.

Just upgrade setuptools and stop using the python3-setuptools Debian
package. Bind it to 41 major version.

Change-Id: I21a0eaa6d16776d0d1d87afb80ecb3677d2a7b94