Standalone Selenium jobs for extensions/skins
Create jobs that run selenium with php7.2/mysql:

    quibble-composer-selenium-docker
    quibble-vendor-selenium-docker

Introduce a new Zuul job template 'extension-quibble-only-selenium' to
trigger those jobs. Intended to be used with
'extension-quibble-noselenium'.

Use vendor repository on master and wmf branches but composer on release
branches.

So for extensions for which it makes sense to extract the Selenium jobs,
we can just replace:

  - extension-quibble
  + extension-quibble-noselenium
  + extension-quibble-only-selenium

Apply that to mediawiki/extensions/Wikibase which would definitely
benefit from it.
Adjust layout test to take in account that 'extension-quibble' might not
be applied on a gated repo.

A similar split has been done for mediawiki/core:
912468a1b32eace4bcd886a0a8b45d83812d2606

Bug: T232759
Change-Id: If7c9440a640ed450054e467ae33e949c3ea3448d