jjb: git plugin does not default to shallow
analytics-refinery-release explicitly asked for a non shallow. That is
the default plugin behavior, there is thus no need to set it explicitly.

Reason for the change is JJB 2.0.10 adds an extra <depth>1</depth>
parameter even when shallowing is disabled. That could a future JJB
upgrade to not show as a noop.

Change-Id: I974cd003c9c7abc92b4282fcb6d4b3c63a99f466