Constrain PyYAML<4
JJB 2.9.1 constrains PyYAML<4 but that is not properly taken in account
by pip/tox. Possibly due to us requiring PyYAML and Zuul requiring at
least 3.1.0.

Dont try to outsmart computer, and just set the same upper constraint
that is set in JJB 2.9.1.

That also remove a warning due to yaml.load(), which in turns causes the
zuul server layout validation to not output anything and cause
utils/zuul-layout-validate.sh to fail due to grep not matching anything.

Ignore grep exist status, it is only for filtering out the spammy zuul
DEBUG and INFO messages. We have set -o pipefail, thus the script would
still fail.

Bug: T236036
Change-Id: I9197cd4047adc5755bf13beac5749c5c9eedbd50