castor: Fallback to loading 'master' cache if ZUUL_BRANCH unset
This is for Zuul 'pipeline' jobs which runs when Git tags are
created (refs/tags/*). There, ZUUL_BRANCH is not set.

For these jobs, the cache will not be saved afterwards (we only
do that for gate-and-submit or postmerge pipelines). This is okay
because there is not much use in sharing cache between tags
and there is no universal way to map it to a closely related branch.

It is also okay if the cache is not so useful for these jobs as
these jobs do not run very often. The main reason we allow this
is to that the same job can easily be re-used between 'postmerge'
and 'publish' pipeline. Thus having castor enabled for the common
case of "master" and "REL1_*" commits, whilst still working
correctly when tags are published as well.

Bug: T232055
Change-Id: Ib39b67255f1f632d76970a9d7e9e6109a799d2b1