Move jjb/diff.sh to utils/jjb-diff.sh
jjb/diff.sh is the sole script in that directory, I keep missing it
since I expect it to be next to the other utilities under ./utils.

Move the file under ./utils with a jjb- prefix.

Update job integration-config-jjb-diff-docker which has to be done in
order to merge this change.  Leaving a symbolic link behind breaks git
rename dectection :/

Have Zuul trigger integration-config-jjb-diff-docker whenever
utils/jjb-diff.sh is modified.

Bug: T236685
Change-Id: I0d1d959b0714c6eac807a52e29062fcb0a9c69ef