Make tox notice dependency changes
When passing dependencies via a file such as:

    deps = -rtest-requirements.txt

Tox is unable to notice a change, it does not include pip to parse the
dependencies. Since the parameter hasn't changed, it would not trigger a
rebuild of the environment.  That is specially harmful when updating
jenkins-job-builder, since people would end up using an old version
unless they notice the venv has to be rebuild.

See upstream issue:
https://github.com/tox-dev/tox/issues/149

Workaround is to explicitly define each of the dependencies and stop
using -r :]

To verify the change, I have used `tox --showconfig --hashseed 0` for
the previous patch then compared with this patch using:

$ colordiff --show-function-line='^\[.*' \
      -u0 config <(tox --hashseed 0 --showconfig)

Which shows the content of files being moved to the deps sections of
each virtualenv. This way if one of the dependency is changed, tox will
notice :]

--- config	2019-10-28 20:22:22.505883073 +0100
+++ newconfig	2019-10-28 20:41:33.289054112 +0100
@@ -95 +95,7 @@ [testenv:zuul_tests]
-  deps            = [-rtest-requirements.txt]
+  deps            = [nose,
+                     PyYAML<4,
+                     python-debian,
+                     git+https://gerrit.wikimedia.org/r/integration/zuul.git#egg=zuul,
+                     python-jenkins==1.1.0,
+                     jenkins-job-builder==3.0.2,
+                     diff-highlight==1.2.0]
@@ -126 +132,3 @@ [testenv:jenkins-jobs]
-  deps            = [-rjenkins-jobs-requirements.txt]
+  deps            = [python-jenkins==1.1.0,
+                     jenkins-job-builder==3.0.2,
+                     diff-highlight==1.2.0]
@@ -157 +165,7 @@ [testenv:qa]
-  deps            = [-rtest-requirements.txt]
+  deps            = [nose,
+                     PyYAML<4,
+                     python-debian,
+                     git+https://gerrit.wikimedia.org/r/integration/zuul.git#egg=zuul,
+                     python-jenkins==1.1.0,
+                     jenkins-job-builder==3.0.2,
+                     diff-highlight==1.2.0]

Bug: T236036
Change-Id: I4f3f8fda20351061cbc3e5af7c6a76cdaf57e574