Stop using zuul-cloner in debian-glue jobs
The debian-glue jobs have to be migrated to Stretch since we are phasing
out Jessie instances.  The jobs rely on zuul-cloner which is provided by
the zuul Debian package. We are not going to package it for Stretch.

Remove zuul-cloner and use raw git commands.  Since the debian-glue jobs
are for a single repository, it is all fine.  The logic is similar to
dockerfiles/ci-src-setup-simple/run.sh.

Update:
    debian-glue
    debian-glue-backports
    debian-glue-backports-non-voting
    debian-glue-buster
    debian-glue-buster-non-voting
    debian-glue-jessie
    debian-glue-jessie-non-voting
    debian-glue-jessie-wikimedia
    debian-glue-jessie-wikimedia-non-voting
    debian-glue-nocheck
    debian-glue-non-voting
    debian-glue-stretch
    debian-glue-stretch-backports
    debian-glue-stretch-backports-non-voting
    debian-glue-stretch-non-voting
    debian-glue-stretch-wikimedia
    debian-glue-stretch-wikimedia-non-voting
    debian-glue-trusty
    debian-glue-trusty-non-voting
    debian-glue-unstable
    debian-glue-unstable-non-voting

Bug: T224943
Change-Id: Id82a1a319639dabd8d6e4c2a931df1121261bcff