Begin migration to pytest, use it as test runner
As the first step to migrating to pytest, begin using it as a test
runner.

The nose attrib plugin is replaced with pytest's mark feature, and tox
commands are updated to use pytest's syntax.

test_coverage uses fixtures instead of the no-longer-used setup()
function.

All yield-style tests were ported to use straight asserts, and some were
converted to use pytest.mark.parametrize where it was trivial.

Change-Id: Icb2a590c65734cff724f9d7a0770ebc58ce4dec8