integration/docroot: Change post-merge message to recommend sudo
Avoid running 'git pull' as your own user because if for whatever
reason files are modified without group-writability, then it is
impossible (or very hard) for someone to fix this up later given
they don't own the file.

This does not solve or prevent the permissions problems of T237707,
but at least it would allow it to be fixed manually by mortals,
because we are all able to act as the doc-uploader user.

Bug: T237707
Change-Id: Ifa89bcb0846a4cfd78a24a088a35c3b5a1440365