train-deploy-notes: Run when wgVersion is changed from -alpha
Instead of looking at ZUUL_REF to determine when a new wmf branch has
been created, detect changes to $wgVersion in DefaultSettings.php. (For
every new wmf branch, we bump the version from an -alpha revision to
-wmf.nn.) If the new value of $wgVersion matches the latest wmf branch,
the normal routine continues, where the notes are generated based on
differences in the latest two wmf branches.

The zuul job definition has also been tweaked to only execute if
includes/DefaultSettings.php is one of the modified files in the merged
patchset. Even so, this new implementation will require slightly more
upfront processing to fetch the ZUUL_REF and inspect the settings file.

Bug: T243330
Change-Id: I786ae8b6e6ca48575dfd094fd23937b1517a95e1