zuul: re-add mediawiki-quibble-… to mediawiki/core#test
Partially reverts d8937a27623. I did not realize at all that
wmf-quibble is *not* a superset of mediawiki-quibble.

The wmf-quibble jobs run *only* tests from extensions.
This is presumably done as optimization so that core tests don't
run on extension commits. In principle that should be fine,
although it does mean that if an extension is breaking core
behaviours we do not catch it anymore. I don't know when that
was decided but I guess we'll need to rely on end-to-end browser
tests and user testing on that front now. PHPUnit integration
tests, despite invoking the service container and extension
registry, are not getting any actual extension intalled in these
jobs when running the core test suites.

Anyway, that's a pre-existing situation. Add back the main
mediawiki-quibble job as otherwise we are not running any
unit tests from MediaWiki core.. on MediaWiki core commits.

(They still ran in the gate, just not on patch submission).

Change-Id: Ie63d9c8ee7e0caefdd0707c213166989b6b0ed34