train-deploy-notes: Get rid of regex anchor
The ref-updated event looks like:

    ...
    "refUpdate": {
        "refName":"refs/heads/wmf/1.35.0-wmf.24",
        "project":"mediawiki/core"
    },
    "type":"ref-updated"
    ...

The "Job.changeMatches" function in zuul looks like:

    def changeMatches(self, change):¬
        matches_branch = False¬
        for branch in self.branches:¬
            if hasattr(change, 'branch') and branch.match(change.branch):
                matches_branch = True¬
            if hasattr(change, 'ref') and branch.match(change.ref):¬
                matches_branch = True¬
        if self.branches and not matches_branch:¬
            return False¬

Since there is no "branch" parameter for the ref-updated change event only a
"ref" parameter the regex for "branch" is actually being matched against the
"ref". Anchoring the regex is causing this job not to get triggered...is my
theory.

Although, manually triggering the "post" pipeline like:

    zuul enqueue --trigger gerrit \
        --pipeline post --project mediawiki/core --change 580341,1

seems to work just fine so this theory could be incorrect.

Bug: T243330
Change-Id: I09ef4ce9e9eb66dd8b08675d5ead1a85ad9731d9