Disable seccomp in Docker
seccomp is a feature of the Linux kernel to add filter on syscalls. So
one can prevent some potentially harmful syscalls entirely.

In October 2019, we have upgraded the CI instances to Stretch which
came with a newer version of libseccomp2. That caused Docker to turn on
seccomp and the containers to run 25/30% slower at least.

The behavior can be easily reproduced locally:

 $ time tox -e flake8 -- --jobs=1
 real	0m1,128s
 user	0m1,065s
 sys	0m0,065s

 $ time firejail --seccomp tox -e flake8 -- --jobs=1
 real	0m2,019s
 user	0m1,861s
 sys	0m0,155s

The zuul_tests env goes from 13 seconds to 23 seconds with seccomp.

One of our heaviest job mwcore-phpunit-coverage-master went from ~ 2h50
to 4h00!

The Docker daemon does not seem to have a configuration setting to
disable seccomp entirely. We thus have to rely on docker run option:
--security-opt seccomp=unconfined

Bug: T236675
Change-Id: If1df87a1f9a170d5eb089dc8e0aaf331824588d8