deploy-notes: Don't checkout core to build notes
Rather than checkout the zuul revision to determine if
"DefaultSettings.php" was modified with a new version of mediawiki,
query gitiles to see if there are changes on the branch that are ahead
of master.

If so, check that this is a change to the latest branch by querying the
gerrit repo for mediawiki core.

If this is the latest branch, rebuild deploy notes.

This will rebuild deploy notes when there are backports to the latest
branch, but, to me, this is a feature not a bug.

Change-Id: Ie0189e9d1ff70ab5a475c4ee777df0a587678dd9