jjb: Let mw-config jobs honour the composer.lock file
Use plain 'composer install' instead of the hacky 'composer-install-dev-only'
that I wrote once. That script was created for testing mediawiki-vendor
in CI where we want to test with using vendor/ for the prod dependencies,
but stil side-load dev dependencies (such as phpunit) on top.

We then re-purposed this for mw-config, because the mw-config repo
had a committed vendor/ directory with wikimedia/cdb, which was still
used in wmf-config at the time.

This in no longer the case, and the vendor/ directory is no longer
commited to Git. Instead, let this job run the normal 'composer install'
command, which has the benefit of honouring the composer.lock file for
recursively pinned dependencies (much like we do with 'npm ci' already).

Affected jobs:
 - operations-mw-config-php72-composer-diffConfig-docker
 - operations-mw-config-php72-composer-lint-docker
 - operations-mw-config-php72-composer-test-docker

Change-Id: Ia387ea6d65eaef5201822304075b697368316bd4