Add pipeline project for TypeScript services
* Add two separate pipelines in a new typescript-service project:
one for test and another for publish.
* Use these for the new push-notifications service.

The main problem solved is to allow the service to use a pipeline config
file so it can customize the stages to write the build output (while the
build can still write to the file system) before `npm run test` is run.

Background: In the mediawiki/services/push-notifications we'd like to
use TypeScript[1]. In the reference patch I created a
.pipeline/config.yaml file but it didn't seem to get used[2].

[1] See files in .pipeline folder of
    If1a0b9baf156731275d57a461473f92798ab1895
[2] integration.wikimedia.org/ci/blue/organizations/jenkins/
service-pipeline-test/detail/service-pipeline-test-and-publish/5527/
pipeline/

Change-Id: Idd20f6e101d3676db1096e3ccd8cec2475d1bb3e