Remove use of Qemu build parameter
The 'parameters' setting in JJB is mostly a joke. It only exists
to create the "Build" web page in the Jenkins UI. None of this
applies when Zuul triggered the job. The 'default' of a parameter
just means that the HTML input field on the "Build" page has this
as the default value.

Zuul submits its own set of parameters, which can only be augmented
by zuul/parameter_functions.py.

Alternatives considered:
* use !include instead of !include-raw. This means the shell script
  would be interpreted as YAML instead of as a simple string,
  and means the .sh file would have to become invalid shell syntax,
  using {{ instead of {, and also the {foo_bar} placeholder would not
  even be escaped for shell purposes, which would make it even more
  fragile.

Bug: T250808
Change-Id: I1e1197edcef91c1a48bfffd0c7d1022ea0d787cc