Schedule publishing sphinx documentation for transferpy
Sphynx documentation for the transfer.py utility (found at repo
software/operations/transferpy) is setup to be built into
doc.wikimedia.org.

After configuring the env to generate the docs:
https://gerrit.wikimedia.org/r/c/operations/software/transferpy/+/602719

For this, I need to schedule a doc build of the master branch
of the repo as a dependency of the copy-to-doc1001 job.

Bug: T253219
Change-Id: I7ae69f13407d396372d22077701b90584f87e46c