jjb: fail docker run command when it is piped
operations-puppet-wmf-style-guide pipes the output with tee in order to
write the job output to a file:

 - docker-run-with-log-cache-src:
     image: docker-registry.wikimedia.org/releng/operations-puppet:0.7.1
     options: '-e RAKE_TARGET=global:wmf_style'
     args: ' | tee "$WORKSPACE/log/rake.log"'

Which is then executed as:

 docker run ... | tee "$WORKSPACE/log/rake.log"

When the build fails, for example due to a git error, the exit code is
shallowed by the pipe and the build is considered as a success. Moreover
since tee creates an empty rake.log file, the warnings plugin would
consider the build to also be a success.

The fix is to use the bash option `pipefail` in order to fail the build
properly. Apply it on the 'docker-run' in case other builds start using
the same dirty hack of piping output.

This change updates pretty much all jobs.

Bug: T253731
Change-Id: I412b4ea0109ad1e52cc4c87ee8618f4b7b3af957