Fix maintenance-disconnect-full-disks definition
It is a pipeline job and most of the config we were setting at the job
level is no more taken in account by the Pipeline plugin. That can be
seen in the job configuration history, Jenkins normalizes the config and
overwrites it as the SYSTEM user resulting in the removal of:

* the 30 minutes build timeout wrapper. The groovy script already has
  a timeout block.
  Update a test to stop checking wether pipeline jobs have the timeout
  wrapper.
* timestamp wrapper. Adding a timestamps {}
* the credentials binding for COMPOSER_GITHUB_OAUTHTOKEN. It is not
  needed by this job anyway.

* <assignedNode>master</assignedNode>
* <canRoam>false</canRoam>

Which are not recognized and already done via node('contint2001') {}

Introduce a new set of defaults 'pipeline-job' which does not have those
wrappers.

This change is a prerequisite for jjb 3.4.0 which make the pipeline jobs
to no more have wrappers generated in the XML config.

Change-Id: Ic24089ff06d29a617d73113aad905cd3d12965ca