dockerfiles: allow fetch when ZUUL_REF=master
An empty repository defaults to point to `master`:

 $ git init .
 $ cat .git/HEAD
 ref: refs/heads/master

Hence when ZUUL_REF is `master`, ci-src-setup.sh bails out cause git
fetch tries to update the current branch. I though the leading + in the
refspec will be enough but the current branch is exempted.

Passing `--update-head-ok` to git fetch lets it update the current
branch, which work down to git 2.1.4 from Jessie.

The issue has been introduced by fcf154f24e9 when changing the refspec
to +$ZUUL_REF:ZUUL_REF.

Rebuild the containers that have that script.

Bug: T256759
Change-Id: I66ae094a5fa54f69aa741fd38670d23be32c917e