Turn on "integrated" phan tests for Parsoid
This tests Parsoid as an extension, which means the extension code is
analyzed in the context of a checked-out copy of mediawiki-core.

Parsoid is in mediawiki/services for historical reasons; we need to add
a couple of special cases to the test suite to let it masquerade as an
extension.

Bug: T248726
Change-Id: Ia07b33ab7df595f3e94d795d50f3e8074fd74ca2