jjb: remove legacy castor job
Delete job castor-save, it is the historical job used to save the
workspace cache. At some point that got duplicated to
castor-save-workspace-cache which was using a Docker container via shell
snippets castor*.bash

Later on castor-save got updated to reuse the same logic resulting in a
mostly duplicate job.  The last jobs still using castor-save were the
ones running on permanent agents which got disposed of as part of
dropping Jessie CI agents (in April 2020).

The credentials 'nodepool-dib-jenkins' were used for the Nodepool
infrastructure and have been removed from the Jenkins master a while

Since castor-save-workspace-cache inherited from the job, inline the
node, build discarding and build parameters. As a result the job is not
modified by this change.

Remove unused jjb/castor-save-sync.bash

Change-Id: Iee3b6727205c2b103c2a2120ba69604c1cd0d76a