Move castor-save-filter.bash check into castor-save-workspacecache.bash
Previously castor-save-filter.bash was used to prevent
castor-save-workspacecache.bash from being executed if ZUUL_PIPELINE
wasn't "gate-and-submit" or "postmerge".  It did this by exiting with
a non-zero status when the required condition was not met.  This
resulted in the job failing even though it is a frequent and normal
situation.

This commit moves the ZUUL_PIPELINE into the top of
castor-save-workspacecache.bash and makes it exit 0 if the required
condition is not met.

Change-Id: I14044496c15b6db78a471175d4916bbf40a7fd21