jjb: debian-glue should not checkout to a branch
jenkins-debian-glue ensure there are local upstream, debian and
pristine-tar branch and attempt to creates them.  In b67e646 I have made
the Jenkins job to checkout the patch to a branch which prevent
debian-glue from deleting it leading to a build failure.

The reason for checking out to a branch was merely to avoid the git
advice to checkout to a branch instead of being in detached mode. Turns
out that can be disabled via git config advice.detachedHead=false.

Stop force checking out to a branch and mute the advice instead.

Jenkins debian glue works just fine when in detached mode, it just just
checkout to a random branch instead.

Update all the debian-glue jobs as well as scap-deb-beta for
consistency.

Bug: T141607
Change-Id: I94f8db3bafbad2cbdfa5a8fda5730d24e92436ac