Revert "Publish coverage reports from extension-codehealth"
This reverts commit 55043fbd4b960f69af8ed29e58ba7058c39a3817.

The CodeHealth pipeline generates PHPUnit code coverage, the intent was
to have it to publish the coverage report. To avoid generating the
coverage twice, we dropped the extension-coverage Zuul template.

The codehealth pipeline oftens generate an empty report, we thus
restored extension-coverage 91d76abb320a.  But there is then a race
condition between the two jobs which might result in an empty report
being published (T242444 / T259109).

This make codehealth pipeline to stop publishing the (empty) HTML
report when it fails and stop the double publishing.

Update mwext-codehealth-master-non-voting

Bug: T242444
Bug: T259109
Change-Id: I726f669c9089694c8d74462ab7edde8eb9e3ecec