dockerfiles: fix tags fetch
Whenever fetching for a `ref-updated` event (such as a new tag), the
$ZUUL_BRANCH parameter is not set which leads to ci-src-setup.sh dieing
out when trying to checkout to that branch. Use the reference name as
the local branch name.

That went broken when implementing the checkout to a branch T141607
since there are no ZUUL_BRANCH set for tags.

Checkout to a detached head instead.

Rebuild the base CI containers, rest will be updated in a follow up
change.

Bug: T186496
Change-Id: I4431a0952673001b01f1d12411e7ea524f858d63