dockerfiles: add test for ci-src-setup.sh
The ci-src-setup.sh script in releng/ci-common is one of the few scripts
that are used all over the place. The recent changes I have made to it
turned out to have glitches which would have been easily discovered with
a proper test suite.

The suite crafts a dummy repository `upstream` to run the script
against.  Execution is done in a `workspace` directory which is wiped
before and after each test to ensure a clean state.

I have added a few tests to notably cover the recent changes made to the
script as well as preexisting features such as GIT_NO_SUBMODULES.

When executing the script, its output is always printed in order to let
nose to capture it and attach it to the output in case there is a test
error.

Instruct tests/test_dockerfiles.py that the `tests` directory is not an
image.

Change-Id: Id405a13d4e4659404bb4b808683bcf50fa0f4067