jjb: support DonationInterface fundraising/REL1_35 branch
DonationInterface does not have a fundraising/REL1_31 branch, it instead
use master which was hardcoded in Quibble with:

  --project-branch mediawiki/extensions/DonationInterface=master

DonationInterface does have a fundraising/REL1_35 branch though, so we
should use it for the REL1_35/php73 job which is only triggered for
fundraising/REL1_35.

A better long term fix would be to normalize DonationInterface to no
more use `master`.

Update quibble-donationinterface-REL1_35-php73-docker

Bug: T261777
Change-Id: I5f2f5ca13ec5574bdf67c0dcbbfaf92351017943