Remove operations-puppet-wmf-style-guide
Was intended to build a nice report of puppet-lint issue. The analysis
has not been running since June and it is unlikely anyone was looking at
the report since it is not advertised.

Maybe later on we can look at crafting an HTML report, publish it
somewhere and send a notification to advertise it.

Meanwhile it serves little purpose. Getting rid of it will also let us
uninstall the Warnings Jenkins plugin which is deprecated.

Bug: T260565
Change-Id: If282e3567bd525a0c8f5450b34426d80141e36bb