jjb: Move fundraising/REL1_35 tests to point to master
We temporarily were using a fundraising/REL1_35 branch of
DonationInterface to hold non-1.31 compatible changes. Now that we
have verified the 1.35 code in production, we are merging that code
into master. This commit leaves the donation_interface_branch var
in place so we can use the same strategy in two years when there's
a new LTS version of core.

Bug: T261777
Change-Id: I21f1c46fc0db1c73dee8cc17f00252d8394b381c