jjb: disable maintenance-sample-workspace-sizes
It runs `du` once per minute on all Jenkins agents which goes against the
disk IO quota that recently has been introduced and regardless adds disk
pressure on the underlying hardware.  The job has its usefulness though
and we might want to reinstate it later, hence just disable it and add a
description linking to the task.

Bug: T266769
Change-Id: I43b9f0f530032c549796a6903ca090586d02e3f0