jjb: use same image in coverage publish job
composer install and phpunit can be run with the same image. They both
ship php 7.2 and we do not need to differentiate between the two steps.
Use the same image for both step, we just have to change the entrypoint
to use php instead of the composer entry point.

A following change will convert the job to a job template and that will
let us pass a single image to the template.

Change-Id: I5cbd95f136eb26c3f97df68fc93b742c163f1292