Extract publishers for analytics and wikidata release builds.
This will reduce duplication and allow easier configuration once we
extract the common templates for all Java projects.

This also creates publisher macros for sending emails to analytics
and discovery mailing lists.

As a side effect, cause analytics-refinery-update-jars-docker to emit an
email on failure and s/GIT_PROJECT/ZUUL_PROJECT/.

Change-Id: I9d79e71d6a1acbc2a1ad4c6bef16f1916d213b56