zuul: make coverage publish meta job appear first
Zuul layout configuration supports setting jobs parameters using a regex
(any job starting with '^') which are known as meta jobs. However if a
job already had parameters applied, the regex is not applied.

Example:

  jobs:
      - name: mwext-phpunit-coverage-docker-publish
        mutex: somemutex
      - name: ^.*-coverage-(docker-)?publish$
        branch: master

Defines the job with a mutex and once it has been defined the regex is
explicitly not matching the already defined job. This result in the job
missing the `branch: master` parameter.

Move the meta job up in the config.  We should probably get all meta
jobs at the very top of the `jobs` config section.

Bug: T270976
Change-Id: I7199c6c4eaa0dd44a32061436095623591c105ab