Allow parsoid to be added to dependency and gatedextensions lists
For historical reasons Parsoid is `mediawiki/services/parsoid` instead
of `mediawiki/extensions/Parsoid`, so names have to be appropriately
remapped.  But for functional purposes it (a) currently acts like an
extension, and (b) will shortly be integrated in core so these
workarounds are temporary.  But at the moment it is very handy to
be able to enable/disable Parsoid on a case-by-case basis by
adding it as an extension dependency.

Change-Id: I3a7038e40b68a16d7f7478fd640fd6582b451f73