Scan fundraising Gerrit projects for PHP vulnerabilities
We do not want to trigger every combinations of (project, branch), they
are filtered using a groovy matrix combination filter to only build the
matching (project, branch).

The matrix-combination filter can be an empty string and JJB happilly
skip it in such a case. It is thus a noop for the other jobs.

Send notifications fundraising tech team in addition of the usual
security team which is notified for all other similar jobs.

Update fundraising-php-security.

Bug: T271352
Change-Id: I104591e80bbc03c85c075b8f0183dc40c15287de