jjb: notify php-security issues from the matrix job
The '{repogroup}-php-security' templates trigger the
'php-composer-security-docker' job which sends an email notification on
failure.  We have a use case to send the notification to a different
list and we cant vary in the job.

Move the logic up to the matrix job which is instructed to notify solely
for the child build failures (not when the matrix job itself fails).

Bug: T271352
Change-Id: Ib4a05bb132b7e3c742dbd1dd9ac779c974f3ba5c