dockerfiles: remove npm npm-stretch and ci-jessie images
Those images were using npm@3.10.10 and used the npm Debian package from
Jessie to bootstrap it. With the upgrade to NodeJS 10, we went with a
npm snapshot in integration/npm.git.

The npm and npm-stretch are thus no more needed and are unused. They are
the last ones descending from ci-jessie so remove that image as well.
Finally!

Bug: T211784
Bug: T224908
Change-Id: I50ff4edc6ed30affe0b8e8fe2d95271c8ca5268f