Remove reference to submodule
This replaces
https://gerrit.wikimedia.org/r/c/integration/config/+/661829 and will
cause our test jobs to fail until we have made the relevant change on
our end:
- merge https://gerrit.wikimedia.org/r/c/wikimedia/fundraising/crm/+/664360

But we can move quickly to have that one merged.

We are decommissioning the drupal repo in favour of the files being
checked into the main repo directly. Hence archive the repository in
Zuul and move it at the bottom of layout.yaml

Bug: T273822
Change-Id: Ia38d9d302d133fd8f1d583a3d15063a75e4d3347