dockerfiles: use pip from distribution
The latest build of tox images cannot execute tox, as pip installed
importlib_metadata v3.4 which requires python 3.6, but system python
is 3.5. The failure in dependency resolution appears to be because we
are on pip 8 and the python_requires statement in setup.cfg is only
validated in pip 9+.

Following back through git history we have been pinned to pip 8.1.2
since the images were running Debian Jessie.  In debian jessie we needed
a newer version of pip than provided by the system to ensure python
wheels were correctly installed. This is no longer a concern with the
images now on stretch, the system ships a sufficiently new version of
pip.

Remove the pip install with explicit version and instead use the Debian
packages.

Bug: T275049
Change-Id: I9a88062f33c7c598ba0adc88e0af26c584969b6c