Add {name}-node10-browser-webdriver-docker
This is a node10-browser-docker version that runs webdriver/chromedriver
when the container is started.

This is an alternative to https://gerrit.wikimedia.org/r/c/wikidata/query/gui/+/666368
which moves the running of webdriver/chromedriver into the code repo
itself.

Right now this only has 1 usage (wikidata/query/gui) but could have more
in the future.

Bug: T222200
Change-Id: Iadfff5db8f1a69dfe7f19d14488e4b25db697f67