jjb: point db dir to the tmpfs in Quibble fullrun jobs
Follow up 6b85c48, it made Docker to spawn the container with a tmpfs
dedicated to the database: `--tmpfs /workspace/db:size=320M`. However
lack of a `--db-dir` parameter passed to Quibble means the database is
initialized in a temporary directory under /tmp.  The patch was thus
useless.

Pass --db-dir to Quibble , which will make the fullrun jobs way faster.

Update:
- integration-quibble-fullrun
- integration-quibble-apache-fullrun

Bug: T266777
Change-Id: Ia69ca00356d5972e828aefbc7de015a3a537f0e6