Add configuration for new wikidata/query-builder repo
This is one step in migrating the wmde/query-builder repository from
GitHub to Gerrit/Jenkins. That becomes advised as part of the security
review process needed to be able to deploy the QueryBuilder in
production in some fashion.

This also gives us the opportunity to get some experience with running
Cypress browser tests on Jenkins.

Together with the changes from
https://github.com/wmde/query-builder/pull/220
this should just work™.

Cypress is now run "inline" with all the other tests and linting checks
done as part of npm test. It needs Chromium and xfvb installed, but more
should not be needed 🤞

Ideas for followups:
* Sonar/CodeHealth for unit-test coverage
* some i18n linting/verification?
* Maybe branch hosting so UX can inspect the patch before it is merged?
* Maybe a post-merge deploy job pushing directly to the Query Builder
  staging environment?

Bug: T277060
Change-Id: Ibd1bbb580d81a41a8de046a60329a6379244bb36