jjb: fix sonar code comparison
CI does checkout the change in a local branch named after the target
branch (typically: `master`), but Sonar then uses that `master` local
branch for comparison and since that is the patch we are scanning, that
yields no difference.

We need to initialize a local branch that will be used as the base to
compare against.

For patch analysis when we use:

 -Dsonar.branch.target=master
 -Dsonar.branch.name=12345,42

CI has the patch for 12345,42 in the master branch.

This change:
* rename the branch holding the patch to 12345-42 ({branch})
* create a `master` which is the parent commit
* check back 12345-42

Change-Id: Ie1bb707f67e8e0f8510c6a6680520affea0923ae