jjb: compress fresnel browser performance reports
Each of the Fresnel test results in json performances report which are
fairly large. The artifacts generated by mediawiki-fresnel-patch-docker
use ~ 25% of the disk space on the Jenkins master and compressing them
would definitely save space.

The only blocker was to keep the possibility to drag'n drop a
trace--trace.json file to Chromium performance report. That has been
done by changing the content-type of file to application/json and set
content-encoding to gzip:
https://gerrit.wikimedia.org/r/c/operations/puppet/+/675515

Update mediawiki-fresnel-patch-docker

Bug: T249268
Change-Id: Ica15705f6c1093379342352dac252ab3c6c229ea